requirements.txt
bleak
fastapi
jinja2
paho-mqtt
python-dotenv
pydantic
uvicorn
